requirements.txt
streamlit
opencv-python
pytesseract
pandas
numpy
pillow